finished case where every cash lost by a user is send to a house bank account

+++ src/main/java/com/mirro/urubudopix/repositories/HouseAccountRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.mirro.urubudopix.models.HouseAccount;
 
 @Repository
-public interface HouseAccountRepository extends JpaRepository<HouseAccount, Long> {
+public interface HouseAccountRepository extends JpaRepository<HouseAccount, Integer> {
 }
